+++ tests/unit/processor/list_comparison/test_list_comparison.py
@@ -1,10 +1,16 @@
 # pylint: disable=missing-docstring
 # pylint: disable=protected-access
 import pytest
+
+import json
+from pathlib import Path
+from unittest import mock
 import responses
 
 from logprep.factory import Factory
 from logprep.processor.base.exceptions import FieldExistsWarning
+from logprep.util.defaults import ENV_NAME_LOGPREP_GETTER_CONFIG
+from logprep.util.getter import HttpGetter
 from tests.unit.processor.base import BaseProcessorTestCase
 
 
@@ -271,11 +277,64 @@ Hans
             "rules": [],
             "list_search_base_path": "http://localhost/tests/testdata/${LOGPREP_LIST}?ref=bla",
         }
+
+        HttpGetter._shared.clear()
+
         processor = Factory.create({"custom_lister": config})
         rule = processor.rule_class.create_from_dict(rule_dict)
         processor._rule_tree.add_rule(rule)
         processor.setup()
         assert processor.rules[0].compare_sets == {"bad_users.list": {"Franz", "Heinz", "Hans"}}
+
+    @responses.activate
+    def test_list_comparison_loads_rule_using_http_and_updates_with_callback(self, tmp_path):
+        target = "localhost/tests/testdata/bad_users.list?ref=bla"
+        url = f"http://{target}"
+        responses.add(
+            responses.GET,
+            url,
+            """Franz
+Heinz
+Hans
+""",
+        )
+        responses.add(
+            responses.GET,
+            url,
+            """Franz
+Heinz
+""",
+        )
+        rule_dict = {
+            "filter": "user",
+            "list_comparison": {
+                "source_fields": ["user"],
+                "target_field": "user_results",
+                "list_file_paths": ["bad_users.list"],
+            },
+            "description": "",
+        }
+        config = {
+            "type": "list_comparison",
+            "rules": [],
+            "list_search_base_path": "http://localhost/tests/testdata/${LOGPREP_LIST}?ref=bla",
+        }
+
+        HttpGetter._shared.clear()
+
+        getter_file_content = {target: {"refresh_interval": 10}}
+        http_getter_conf: Path = tmp_path / "http_getter.json"
+        http_getter_conf.write_text(json.dumps(getter_file_content))
+        mock_env = {ENV_NAME_LOGPREP_GETTER_CONFIG: str(http_getter_conf)}
+        with mock.patch.dict("os.environ", mock_env):
+            processor = Factory.create({"custom_lister": config})
+            rule = processor.rule_class.create_from_dict(rule_dict)
+            processor._rule_tree.add_rule(rule)
+            processor.setup()
+            assert processor.rules[0].compare_sets == {"bad_users.list": {"Franz", "Heinz", "Hans"}}
+            assert processor.rules[0].compare_sets == {"bad_users.list": {"Franz", "Heinz", "Hans"}}
+            HttpGetter(target=target, protocol="http").scheduler.run_all()
+            assert processor.rules[0].compare_sets == {"bad_users.list": {"Franz", "Heinz"}}
 
     def test_list_comparison_does_not_add_duplicates_from_list_source(self):
         document = {"users": ["Franz", "Alpha"]}